Update Item.js
I think for the incident date, we only need to get the date part of it in YYYYMMDD format. This will be easier to sort in future instances.

If we want to revert back to the full date, we can update the schema in the future to

incidentDate: {type: Date, required: true, default: Date.now}