Added redux
The problem here is that with the file componentDidMount() {
this.props.getItems();
}

in the IncidentList.js will cause an error

"Error: You may not call store.getState() while the reducer is executing. The reducer has already received the state as an argument. Pass it down from the top reducer instead of reading it from the store."